Fixed and added
Fixed the navbar to display highlighted icon on each page, appointment access from chats is broken however. Also changed the name of appointment.dart and to appointment_page.dart since the name was a duplicate and confusing between the object and the actual page. Lastly moved 2 of the files, patientlist and appointment into the pages folder.
