+++ lib/pages/dashboard.dart
@@ -7,7 +7,7 @@ import 'package:medipal/main.dart';
 import 'package:firebase_database/firebase_database.dart';
 import 'package:medipal/objects/patient.dart';
 import 'package:medipal/pages/patientpage.dart';
-import 'package:medipal/patient_list.dart';
+import 'package:medipal/pages/patient_list.dart';
 import 'package:flutter/material.dart';
 import 'package:medipal/patient_form.dart';
 import 'package:medipal/objects/patient.dart';
@@ -15,7 +15,7 @@ import 'package:medipal/patient_data.dart';
 import 'package:flutter/material.dart';
 import 'package:medipal/constant/images.dart';
 import 'package:medipal/pages/appointment_date.dart';
-import 'package:medipal/pages/appointment.dart';
+import 'package:medipal/pages/appointment_page.dart';
 
 // from appointment_data.dart will be replaced with object/appointment.dart
 class Appointment {
